requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-wtf==1.1.1
wtforms==3.0.1
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
